Fix a crash caused by dynamic variables
When this provider is used and another resource is providing values for a resource managed by this provider, attributeHash() might be given nil. Instead of panicking on the type assertion, hash the "nil" value instead. After the plan runs, the values are known and properly submitted to LDAP.

A particular use case: when managing VMs though terraform-provider-libvirt, and managing DHCPd through LDAP using this provider. When a machine is created, a random MAC address is generated. This same MAC address is needed to properly instruct DHCPd to give out a fixed IP and thus using terraform, it can be queried from the libvirt resource. The alternative is to hardcode the MAC addresses but this should not be necessary at al.